backend/requirements.txt
alembic==1.7.7
apispec==5.1.1
click==8.0.4
Flask==2.0.3
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
flask-smorest==0.37.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.11.3
importlib-resources==5.6.0
itsdangerous==2.1.2
Jinja2==3.1.1
Mako==1.2.0
MarkupSafe==2.1.1
marshmallow==3.15.0
marshmallow-sqlalchemy==0.28.0
packaging==21.3
pyparsing==3.0.7
python-dotenv==0.20.0
six==1.16.0
SQLAlchemy==1.4.32
webargs==8.1.0
Werkzeug==2.0.3
zipp==3.7.0
